test(dedupe): consolidate request_validator tests (5 → 2 files)
Consolidate RequestValidator test files in tests/unit/gpt_trader/security:
- test_request_validator_success.py + invalid_action.py + missing_fields.py + edge_cases.py → test_request_validator.py
- test_request_validator_size_limits.py → test_request_validator_limits.py (rename)

Cluster: ba5346200fdc
Files: 5 → 2 (-3)
Tests: 18 passing
