+++ tests/_triage/dedupe_candidates.yaml
@@ -2467,7 +2467,9 @@ clusters:
     expected_file_delta: -4
     priority: low
     rationale: Consolidate 5 test files into single parametrized file
-    status: pending
+    status: done
+    owner: claude
+    pr_url: https://github.com/Solders-Girdles/GPT-Trader/pull/315
     added: '2026-01-20'
   bcf5f1feb7e9:
     type: similar_names
